// src/app/models/hospitalSignupModel.Interface.ts
export interface HospitalSignupModel {
    Id?: number;
    BusinessName?: string;
    CityId?: number;
    City?: any;
    Address1?: string;
    Address2?: string;
    Lat?: string;
    Long?: string;
    PhoneNumber?: string;
    Email?: string;
    ProfilePic?: string;
    AboutMe?: string;
    Link1?: string;
    Link2?: string;
    Status?:number;
    OfficeTimings?: [];
}